+++ Panther/include/diagnostics.h
@@ -92,8 +92,9 @@ namespace pcit::panther{
 		SemaInvalidDiscardStmtRHS,
 		SemaInvalidAddrOfRHS,
 		SemaInvalidDerefRHS,
-		SemaInvalidAsLHS,
-		SemaInvalidAsRHS,
+		SemaInvalidInfixLHS,
+		SemaInvalidInfixRHS,
+		SemaInvalidInfixArgTypes,
 		SemaInvalidTypeQualifiers,
 		SemaParamTypeVoid,
 		SemaNoMatchingFunction,
@@ -213,8 +214,9 @@ namespace pcit::panther{
 			break; case DiagnosticCode::SemaInvalidDiscardStmtRHS:                return "S";
 			break; case DiagnosticCode::SemaInvalidAddrOfRHS:                     return "S";
 			break; case DiagnosticCode::SemaInvalidDerefRHS:                      return "S";
-			break; case DiagnosticCode::SemaInvalidAsLHS:                         return "S";
-			break; case DiagnosticCode::SemaInvalidAsRHS:                         return "S";
+			break; case DiagnosticCode::SemaInvalidInfixLHS:                      return "S";
+			break; case DiagnosticCode::SemaInvalidInfixRHS:                      return "S";
+			break; case DiagnosticCode::SemaInvalidInfixArgTypes:                 return "S";
 			break; case DiagnosticCode::SemaInvalidTypeQualifiers:                return "S";
 			break; case DiagnosticCode::SemaParamTypeVoid:                        return "S";
 			break; case DiagnosticCode::SemaNoMatchingFunction:                   return "S";